refactor(ui): break the flow store / useDependencies import cycle
stores/flow.ts imported transformResponse from the useDependencies
composable, which imports useFlowStore back. The cycle also put the whole
composable -- five stores, i18n, the design-system barrel -- in the static
closure of every page reaching the flow store, the login form included.

transformResponse is a pure function, so move it to a leaf module beside
the types it uses. The login chunk now holds no dashboard component.
